fix type for running sum

+++ flink-analyzer/src/main/java/operators/aggregate/AverageAccumulator.java
@@ -3,6 +3,6 @@ package operators.aggregate;
 // The accumulator, which holds the state of the in-flight aggregate
 public class AverageAccumulator {
     long count;
-    long sum;
+    double sum;
 }
